Dispose registrations
