+++ src/libraries/System.Net.Quic/src/System.Net.Quic.csproj
@@ -9,6 +9,7 @@
     <!-- All configurations -->
     <Compile Include="System\Net\Quic\Implementations\MsQuic\Internal\MsQuicAddressHelpers.cs" />
     <Compile Include="System\Net\Quic\Implementations\MsQuic\Internal\MsQuicApi.cs" />
+    <Compile Include="System\Net\Quic\Implementations\MsQuic\Internal\MsQuicParameterHelpers.cs" />
     <Compile Include="System\Net\Quic\Implementations\MsQuic\Internal\MsQuicSecurityConfig.cs" />
     <Compile Include="System\Net\Quic\Implementations\MsQuic\Internal\MsQuicSession.cs" />
     <Compile Include="System\Net\Quic\Implementations\MsQuic\Internal\MsQuicStatusException.cs" />
