+++ src/libraries/System.Net.Quic/src/System/Net/Quic/Implementations/MsQuic/Internal/MsQuicAddressHelpers.cs
@@ -1,30 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
 using System.Net.Sockets;
-using System.Text;
-using System.Threading.Tasks;
 using static System.Net.Quic.Implementations.MsQuic.Internal.MsQuicNativeMethods;
 
 namespace System.Net.Quic.Implementations.MsQuic.Internal
 {
-    internal static class MsQuicParameterHelpers
-    {
-        internal static unsafe SOCKADDR_INET GetINetParam(MsQuicApi api, IntPtr nativeObject, uint level, uint param)
-        {
-            byte* ptr = stackalloc byte[sizeof(SOCKADDR_INET)];
-            QuicBuffer buffer = new QuicBuffer
-            {
-                Length = (uint)sizeof(SOCKADDR_INET),
-                Buffer = ptr
-            };
-
-            MsQuicStatusException.ThrowIfFailed(api.UnsafeGetParam(nativeObject, level, param, ref buffer));
-
-            return *(SOCKADDR_INET*)ptr;
-        }
-    }
-
     internal static class MsQuicAddressHelpers
     {
         internal const ushort IPv4 = 2;
